+++ packages/shared/src/index.ts
@@ -1,4 +1,6 @@
+export * from './env'
 export * from './locale'
+export * from './project'
 export * from './changelog'
 export * from './roadmap'
 export * from './services'